<fix> 포스트 리스트를 보낼때 쓰는 용도로 msg가 붙어있지 않은 단일포스트의DTO 생성, 삭제에서 패스워드 검증 추가

+++ src/main/java/com/homework/springhomework/service/PostService.java
@@ -1,9 +1,6 @@
 package com.homework.springhomework.service;
 
-import com.homework.springhomework.dto.DeleteResponseDto;
-import com.homework.springhomework.dto.RequestDto;
-import com.homework.springhomework.dto.ResponsePostDto;
-import com.homework.springhomework.dto.ResponseTotalPostDto;
+import com.homework.springhomework.dto.*;
 import com.homework.springhomework.entity.Post;
 import com.homework.springhomework.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
@@ -27,14 +24,14 @@ public class PostService {
 
 
     @Transactional(readOnly = true) // 전체 포스트 보여주기
-    public ResponseTotalPostDto getPostList() { // 얘가 안오는 문제가 있는 부분이구요..
+    public ResponseTotalPostDto getPostList() {
         List<Post> list = postRepository.findAllByOrderByModifiedAtDesc();
-        List<ResponsePostDto> postsResponseDtoList = new ArrayList<>();
+        List<ResponsePostNoMsgDto> responsePostNoMsgDtoList = new ArrayList<>();
         for (Post post : list) {
-            ResponsePostDto responsePostDto = new ResponsePostDto(post);
-            postsResponseDtoList.add(responsePostDto);
+            ResponsePostNoMsgDto responsePostNoMsgDto = new ResponsePostNoMsgDto(post);
+            responsePostNoMsgDtoList.add(responsePostNoMsgDto);
         }
-        ResponseTotalPostDto responseTotalPostDto = new ResponseTotalPostDto(postsResponseDtoList, "진심 스스로불러온 재앙이구요. 스스로 작성한 API 명세서에 고통받는다");
+        ResponseTotalPostDto responseTotalPostDto = new ResponseTotalPostDto(responsePostNoMsgDtoList, "포스트정보에 msg가 하나씩 붙은 100개의 dto가 든 리스트에 메시지 하나 더붙어 가는 꼴을 면하기 위한 몸부림");
 
         return responseTotalPostDto;
     }
@@ -42,7 +39,7 @@ public class PostService {
     @Transactional(readOnly = true) //특정 단일 포스트 보여주기
     public ResponsePostDto getPost(Long id) {
         Post post = postRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("id값이 일치하는 포스트 없음!"));
-        ResponsePostDto responsePostDto = new ResponsePostDto(post, "msg를 굳이 넣지 않아도 되는 이유 너무나 절실하게 깨우침...");
+        ResponsePostDto responsePostDto = new ResponsePostDto(post, "이것은 id값으로 특정 단일 포스트 보여주기");
 
         return responsePostDto;
     }
@@ -61,29 +58,17 @@ public class PostService {
     }
 
     @Transactional
-    public DeleteResponseDto deletePost(Long id) {
-        postRepository.deleteById(id);
-        DeleteResponseDto deleteResponseDto = new DeleteResponseDto("삭제완료");
+    public DeleteResponseDto deletePost(Long id, String password) {
+        Post post = postRepository.findById(id).orElseThrow(() -> new RuntimeException("해당 id의 포스트가 존재하지 않습니다."));
+        if (post.checkPw(password)) {
+            postRepository.deleteById(id);
+        } else throw new RuntimeException("비밀번호가 틀렸습니다.");
+        DeleteResponseDto deleteResponseDto = new DeleteResponseDto("포스트가 정상적으로 삭제되었습니다.");
         return deleteResponseDto;
     }
 
-
-//    public Long update(Long id, PostRequestDto postRequestDto) {
-//        Post post = postRepository.findById(id).orElseThrow(
-//                () -> new RuntimeException(("해당 id의 포스트가 존재하지 않습니다"))
-//        );
-//        String pw = postRequestDto.getPassword(); // 입력한 비번
-//        String correctPw = post.getPassword(); // 업데이트할 포스트가 갖고있는 비번
-//
-//        if(pw.equals(correctPw)){
-//            // 수정 작업 로직이 여기서 돌아가.
-//            return post.getId(); // 리턴을 왜 시키지...?? x 415번 :이유 = 작동에 필요한거 아님. 걍 API에 이걸 뱉으라고 되있어서.(보통 실무에선 프론트에서 필요해서.나는 여기선 API에 맞춰서 필요한거 뱉거나 필요없으면 VOID 해도 됨)
-//        } else {
-//            throw new RuntimeException("비밀번호가 맞지 않습니다.");
-
-
 } // 클래스 PostService 끝
-
+// 아래는 고민했던 흔적이라서 남김
 
 //    @Transactional(readOnly = true) //전체 포스트 조회
 //    public List<Post> getPosts() {return postRepository.findAllByOrderByModifiedAtDesc();}
